requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.1.0
stripe==7.5.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-filter==23.3
whitenoise==6.7.0